+++ R/ggcorrmat.R
@@ -19,22 +19,12 @@
 #'   literally anything other than `"plot"`), a dataframe containing all details
 #'   from statistical analyses (e.g., correlation coefficients, statistic
 #'   values, *p*-values, no. of observations, etc.) will be returned.
-#' @param matrix.type Character, `"full"` (default), `"upper"` or `"lower"`,
+#' @param matrix.type Character, `"lower"` (default), `"full"`, or `"upper"`,
 #'   display full matrix, lower triangular or upper triangular matrix.
-#' @param matrix.method The visualization method of correlation matrix to be
-#'   used. Allowed values are `"square"` (default) or `"circle"`.
 #' @param sig.level Significance level (Default: `0.05`). If the *p*-value in
 #'   *p*-value matrix is bigger than `sig.level`, then the corresponding
 #'   correlation coefficient is regarded as insignificant and flagged as such in
 #'   the plot. Relevant only when `output = "plot"`.
-#' @param p.adjust.method What adjustment for multiple tests should be used?
-#'   (`"holm"`, `"hochberg"`, `"hommel"`, `"bonferroni"`, `"BH"`, `"BY"`,
-#'   `"fdr"`, `"none"`). See `stats::p.adjust` for details about why to use
-#'   `"holm"` rather than `"bonferroni"`). Default is `"none"`. If adjusted
-#'   *p*-values are displayed in the visualization of correlation matrix, the
-#'   **adjusted** *p*-values will be used for the **upper** triangle, while
-#'   **unadjusted** *p*-values will be used for the **lower** triangle of the
-#'   matrix.
 #' @param colors A vector of 3 colors for low, mid, and high correlation values.
 #'   If set to `NULL`, manual specification of colors will be turned off and 3
 #'   colors from the specified `palette` from `package` will be selected.
@@ -42,9 +32,9 @@
 #'   coefficients (only valid when `insig = "pch"`). Default: `pch = "cross"`.
 #' @param ggcorrplot.args A list of additional (mostly aesthetic) arguments that
 #'   will be passed to `ggcorrplot::ggcorrplot` function. The list should avoid
-#'   any of the following arguments since they are already internally being used
-#'   by `ggstatsplot`: `corr`, `method`, `p.mat`, `sig.level`, `ggtheme`,
-#'   `colors`, `matrix.type`, `lab`, `pch`, `legend.title`, `digits`.
+#'   any of the following arguments since they are already internally being
+#'   used: `corr`, `method`, `p.mat`, `sig.level`, `ggtheme`, `colors`, `lab`,
+#'   `pch`, `legend.title`, `digits`.
 #' @inheritParams statsExpressions::expr_corr_test
 #' @inheritParams ggbetweenstats
 #' @inheritParams theme_ggstatsplot
@@ -54,7 +44,7 @@
 #' @import ggplot2
 #'
 #' @importFrom ggcorrplot ggcorrplot
-#' @importFrom dplyr select
+#' @importFrom dplyr select matches
 #' @importFrom purrr is_bare_numeric keep
 #' @importFrom rlang !! enquo quo_name is_null
 #' @importFrom ipmisc green blue yellow red
@@ -100,17 +90,16 @@ ggcorrmat <- function(data,
                       cor.vars = NULL,
                       cor.vars.names = NULL,
                       output = "plot",
-                      matrix.type = "full",
-                      matrix.method = "square",
+                      matrix.type = "lower",
                       type = "parametric",
                       beta = 0.1,
                       k = 2L,
                       sig.level = 0.05,
                       conf.level = 0.95,
                       bf.prior = 0.707,
-                      p.adjust.method = "none",
+                      p.adjust.method = "holm",
                       pch = "cross",
-                      ggcorrplot.args = list(outline.color = "black"),
+                      ggcorrplot.args = list(method = "square", outline.color = "black"),
                       package = "RColorBrewer",
                       palette = "Dark2",
                       colors = c("#E69F00", "white", "#009E73"),
@@ -197,15 +186,8 @@ ggcorrmat <- function(data,
   # ========================== plot =========================================
 
   # create matrices for correlation coefficients and p-values
-  corr.mat <-
-    df_correlation %>%
-    dplyr::select(dplyr::matches("^parameter|^r")) %>%
-    as.matrix()
-
-  p.mat <-
-    df_correlation %>%
-    dplyr::select(dplyr::matches("^parameter|^p")) %>%
-    as.matrix()
+  corr.mat <- as.matrix(dplyr::select(df_correlation, dplyr::matches("^parameter|^r")))
+  p.mat <- as.matrix(dplyr::select(df_correlation, dplyr::matches("^parameter|^p")))
 
   # creating the basic plot
   # if user has not specified colors, then use a color palette
@@ -264,7 +246,6 @@ ggcorrmat <- function(data,
     rlang::exec(
       .f = ggcorrplot::ggcorrplot,
       corr = corr.mat,
-      method = matrix.method,
       p.mat = p.mat,
       sig.level = sig.level,
       ggtheme = ggtheme,
@@ -321,5 +302,5 @@ ggcorrmat <- function(data,
 
   # if any additional modification needs to be made to the plot
   # this is primarily useful for grouped_ variant of this function
-  return(plot + ggplot.component)
+  plot + ggplot.component
 }